Display.tr: convert dimensions to Float64.

+++ picar_lidar/viewing.mojo
@@ -28,7 +28,7 @@ struct Display:
         self.surface.fill(self.pygame.Color("white"))
 
     fn tr(self, point: Point) -> Point:
-        return Point(point[0] + self.width / 2, point[1] + self.height / 2)
+        return Point(point[0] + Float64(self.width) / 2.0, point[1] + Float64(self.height) / 2.0)
 
     fn draw_lines(self, color: String, lines: List[Line]) raises:
         var x1: Float64
